server/requirements.txt
flask
configparser
flask_mysqldb
pandas
matplotlib
requests
PyMySQL
mysql.connector
sklearn
scikit-learn
